+++ Changelog.md
@@ -5,6 +5,10 @@
 1.  Remove TinyXML symbols from public API: Move uninstalled headers
     * [Pull request 662](https://bitbucket.org/osrf/sdformat/pull-requests/662)
 
+1. parser addNestedModel: check `//axis/xyz/@expressed_in` before rotating joint axis.
+    * [Pull request 657](https://bitbucket.org/osrf/sdformat/pull-requests/657)
+    * [Issue 219](https://bitbucket.org/osrf/sdformat/issues/219)
+
 ### SDFormat 9.1.0 (2020-01-29)
 
 1. Remove URI scheme, if present, when finding files.